session models moved to sessions/model package . Some changes applied . Not SessionEntity is just a template not real entity . Validation messages fixed
