--- messenger-core/src/main/java/com/example/messenger/session/SessionDeleteDto.java
@@ -1,25 +1,0 @@
-package com.example.messenger.session;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.Accessors;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-/**
- * DTO for {@link SessionEntity}
- */
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Accessors(chain = true)
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class SessionDeleteDto implements Serializable {
-    @NotNull(message = "Session Id cannot be null")
-    private String user1;
-    private String user2;
-}
