+++ messenger-core/src/main/java/com/example/messenger/session/model/SessionDto.java
@@ -1,0 +1,14 @@
+package com.example.messenger.session.model;
+
+import java.util.UUID;
+
+/**
+ * DTO for {@link SessionEntity}
+ */
+public record SessionDto(
+
+        UUID id,
+
+        String user1,
+
+        String user2) {}
