+++ messenger-core/src/main/java/com/example/messenger/session/model/SessionMapper.java
@@ -1,9 +1,11 @@
-package com.example.messenger.session;
+package com.example.messenger.session.model;
 
 import org.mapstruct.*;
 import org.mapstruct.factory.Mappers;
 
-@Mapper
+@Mapper(unmappedSourcePolicy = ReportingPolicy.IGNORE,
+        unmappedTargetPolicy = ReportingPolicy.IGNORE,
+        componentModel = MappingConstants.ComponentModel.SPRING)
 public interface SessionMapper {
 
     SessionMapper INSTANCE = Mappers.getMapper(SessionMapper.class);
