+++ messenger-core/src/main/java/com/example/messenger/session/service/SessionServiceImpl.java
@@ -5,13 +5,16 @@ import com.example.messenger.exceptions.CustomEntityNotFoundException;
 import com.example.messenger.exceptions.CustomValidationException;
 import com.example.messenger.exceptions.EntityAlreadyExistException;
 import com.example.messenger.kafka.CommandProducer;
-import com.example.messenger.message.MessageEntity;
-import com.example.messenger.message.MessageMapper;
 import com.example.messenger.message.MessageService;
 import com.example.messenger.user.UserDto;
 import com.example.messenger.user.UserMapper;
 import com.example.messenger.user.UserService;
 import com.example.messenger.user.UserUpdateDto;
+import com.example.messenger.session.model.*;
+import com.example.messenger.session.query.SessionQueryClient;
+import com.example.messenger.user.model.UserEntity;
+import com.example.messenger.user.service.UserService;
+import com.example.messenger.utility.AuthorityUpdateType;
 import com.example.messengerutilities.utility.DataTypes;
 import com.example.messengerutilities.utility.RequestTypes;
 import com.example.messengerutilities.utility.TopicNames;
